# Allow non UTF8 characters in message