--- a/go.mod
+++ b/go.mod
@@ -5,6 +5,7 @@ require (
 	github.com/leodido/ragel-machinery v0.0.0-20181214104525-299bdde78165
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/stretchr/testify v1.2.2
+	golang.org/x/text v0.3.2
 )
 
 go 1.13