--- a/rfc5424/machine.go.rl
+++ b/rfc5424/machine.go.rl
@@ -38,8 +38,8 @@ const (
 	ErrSdParam        = "expecting a structured data parameter (`key=\"value\"`, both part from 1 to max 32 US-ASCII characters; key cannot contain `=`, ` `, `]`, and `\"`, while value cannot contain `]`, backslash, and `\"` unless escaped)"
 	// ErrMsg represents an error in the MESSAGE part of the RFC5424 syslog message.
 	ErrMsg            = "expecting a free-form optional message in UTF-8 (starting with or without BOM)"
-	// ErrMsgNonUTF8 represents an error in the MESSAGE part of the RFC5424 syslog message if AllowNonUTF8InMessage is enabled.
-	ErrMsgNonUTF8     = ErrMsg + " or a free-form optional message in any encoding (starting without BOM)"
+	// ErrMsgNotCompliant represents an error in the MESSAGE part of the RFC5424 syslog message if WithCompliatMsg option is on.
+	ErrMsgNotCompliant     = ErrMsg + " or a free-form optional message in any encoding (starting without BOM)"
 	// ErrEscape represents the error for a RFC5424 syslog message occurring when a STRUCTURED DATA PARAM value contains '"', '\', or ']' not escaped.
 	ErrEscape         = "expecting chars `]`, `\"`, and `\\` to be escaped within param value"
 	// ErrParse represents a general parsing error for a RFC5424 syslog message.
@@ -65,10 +65,10 @@ action markmsg {
 	m.msgat = m.p
 }
 
-action choose_msg_encoding {
+action select_msg_mode {
 	fhold;
 
-	if m.allowNonUTF8InMessage {
+	if m.compliantMsg {
 		fgoto msg_any;
 	}
 	fgoto msg_utf8;
@@ -239,8 +239,8 @@ action err_msg {
 		output.message = string(m.data[m.msgat:m.p])
 	}
 
-	if m.allowNonUTF8InMessage {
-		m.err = fmt.Errorf(ErrMsgNonUTF8 + ColumnPositionTemplate, m.p)
+	if m.compliantMsg {
+		m.err = fmt.Errorf(ErrMsgNotCompliant + ColumnPositionTemplate, m.p)
 	} else {
 		m.err = fmt.Errorf(ErrMsg + ColumnPositionTemplate, m.p)
 	}
@@ -305,7 +305,7 @@ msg_utf8 := (bom? utf8octets) >mark >markmsg %set_msg $err(err_msg);
 # MSG = MSG-ANY | MSG-UTF8
 msg_any := ((bom utf8octets) | (any* - (bom any*))) >mark >markmsg %set_msg $err(err_msg);
 
-msg = any? @choose_msg_encoding;
+msg = any? @select_msg_mode;
 
 fail := (any - [\n\r])* @err{ fgoto main; };
 
@@ -326,7 +326,7 @@ type machine struct {
 	msgat        int
 	backslashat  []int
 	bestEffort 	 bool
-	allowNonUTF8InMessage bool
+	compliantMsg bool
 }
 
 // NewMachine creates a new FSM able to parse RFC5424 syslog messages.