--- a/rfc5424/options.go
+++ b/rfc5424/options.go
@@ -12,10 +12,18 @@ func WithBestEffort() syslog.MachineOption {
 	}
 }
 
-// AllowNonUTF8InMessage allows non utf8 characters in the message part.
-func AllowNonUTF8InMessage() syslog.MachineOption {
+// WithCompliantMsg enables the parsing of the MSG part of the Syslog as per RFC5424.
+//
+// When this is on, the MSG can either be:
+// - an UTF-8 string which starts with a BOM marker
+// or
+// - a free-form message (0-255) not starting with a BOM marker.
+//
+// Ref.: https://tools.ietf.org/html/rfc5424#section-6.4
+// Ref.: https://tools.ietf.org/html/rfc5424#section-6
+func WithCompliantMsg() syslog.MachineOption {
 	return func(m syslog.Machine) syslog.Machine {
-		m.(*machine).allowNonUTF8InMessage = true
+		m.(*machine).compliantMsg = true
 		return m
 	}
 }