… README.md
… docs/rfc5424_msg_any.dot
… docs/rfc5424_msg_compliant.dot
… docs/rfc5424_msg_utf8.dot
… go.sum