@bastjan 2020-05-01
Thanks for the review and the refinements!

I think the name `WithCompliantMsg` reflects the intended option better and I'm all for it.

> Since the default behavior (no option passed in) stays as is, if we want, we could also relax it more, which should lead to even better performances. In this case a lot of test cases should be updated, tho. :)

So the default (no option) would be relaxed to accept any message; `WithCompliantMsg` would make the parser more strict?