@bastjan 2020-05-01
Sounds good to me! I could tackle updating the test cases over the next two days if you need any help.