@leodido 2020-05-01
I can proceed right now with the option rename.

But, I'd love if you could help me with the second and last step of this PR (relaxing the default behaviour and refactoring the test cases)