@bastjan 2020-05-01
> But, I'd love if you could help me with the second and last step of this PR (relaxing the default behaviour and refactoring the test cases)

Sure, glad to help :)