@bastjan 2020-05-03
Hey @leodido,

I relaxed the default parsing, updated the tests (there are many 😅), and rerun the benchmarks in the PR description. Mind to take a look? :)

I also relaxed the builder to not create a situation where a parser->builder roundtrip would remove the MSG. Is this alright with you?