@leodido 2020-05-10
Hey @bastjan yes, relaxing the builder was needed.

Ideally, in a second PR, we should approach the same topic for builder.

Ie., making the builder configurable (compliant message or not) and its `SetMessage` to work accordingly. There are already (empty) builder test cases for such scenarios. 

Anyways, you did a great job, thanks for approaching issue #21 .. I'm about to push some little final refinements and then merge this 🎉